Manage data Model submit new expense in both component
