+++ src/app/dashboard/expenses/add-expense-dialog/add-expense-dialog.component.html
@@ -36,23 +36,40 @@
       <label class="block text-sm font-medium mb-1">Category</label>
       <p-select formControlName="selectedCategory" [options]="categories" optionLabel="category" optionValue="id" placeholder="Select" class="w-full md:w-56" />
     </div>
-    <!-- <div class="">
+    <div class="">
       <label class="block text-sm font-medium mb-1">Split Type</label>
       <p-selectButton 
         [options]="splitOptions" 
-        formControlName="spitType" 
+        formControlName="splitType" 
         optionLabel="label" 
         optionValue="value" />
-    </div> -->
+    </div>
     <p-divider class="m-0"/>
     
     <div>
-      <label class="block text-sm font-medium mb-1">PaidBy</label>
-      <p-select formControlName="paidBy" [options]="groupMembers" optionLabel="name" optionValue="id" placeholder="Member" class="w-full md:w-56" />
-    </div>
+      <label class="block text-sm font-medium mb-1">Members</label>
+      <p-multiSelect
+        [options]="groupMembers"
+        formControlName="selectedMembers"
+        optionLabel="name"
+        display="chip"
+        [showToggleAll]="false"
+        [filter]="true"
+        [showHeader]="false"
+        [selectedItemsLabel]="'{0} selected'"
+        [panelStyle]="{ 'min-width': '220px', 'max-height': '210px', 'overflow-y': 'auto' }"
+        class="w-full"
+        [showClear]="true"
+        [virtualScroll]="true"
+        [virtualScrollItemSize]="38"
+        checkbox="true"
+        [appendTo]="'body'"
+        placeholder="Select Members"
+      ></p-multiSelect>
+      </div>
     <div class="flex justify-end gap-2 mt-4">
       <button pButton type="button" label="Cancel" class="p-button-text" (click)="onClose()"></button>
-      <button pButton type="submit" label="Add Expense" class="p-button-primary" [disabled]="newExpense.invalid"></button>
+      <button pButton type="submit" label="Add Expense" class="p-button-primary"></button>
     </div>
   </form>
 </p-dialog>
